+++ lib/tans-parser.rb
@@ -7,4 +7,6 @@ end
 require_relative "tans_parser/version"
 require_relative "tans_parser/ansi_parser"
 require_relative "tans_parser/ansi_utils"
+require_relative "tans_parser/element"
+require_relative "tans_parser/selector"
 require_relative "tans_parser/state"